AV2

+++ Atividades Ajalmar/AV2(2022.2)/pet.c
@@ -37,6 +37,23 @@ int checkCodUser(char *file_path,int TestCodUser){
     }
     return 1;
 }
+int checkCodPet(char *file_path,int TestCodPet){
+
+    int codPet=0;
+    FILE *chv=NULL;
+    chv=fopen(file_path,"rb");
+
+    fseek(chv,0,SEEK_SET);
+    fread(&codPet,sizeof(int),1,chv);
+    printf("%d",codPet);
+    for(int i=0;i<codPet;i++){
+
+        if(TestCodPet==i+1){
+            return 0;
+        }
+    }
+    return 1;
+}
 void insertPetsUI() {
     int codUser, option;
     char *Name = malloc(sizeof(char) * sizeMaxNames);
@@ -113,9 +130,9 @@ void insertPets(char *file_path,char *Name,char *Date,char *Type,int codUser){
 
 }
 
-int deletePerson(char *file_path,int codPet){
+int deletePets(char *file_path,int codPet){
     int option;
-    if(checkCodUser("persons.bin",codUser)){
+    if(!checkCodPet("keyPets.bin",codPet)){
         printf("\n#---------------------------------------------------#");
         printf("\n|               USUARIO INEXISTENTE                 |");
         printf("\n#---------------------------------------------------#");
@@ -126,23 +143,24 @@ int deletePerson(char *file_path,int codPet){
             printf("\n > ");
             scanf("%d", &option);
             if (option == 1)
-                interfaceUser();
+                interfacePets();
         } while (option != 1);
     }
-    codUser--;
+    codPet--;
     FILE *archive=NULL;
     archive=fopen(file_path,"rb+");
 
     if(archive==NULL)
         archive=fopen(file_path,"wb+");
-    person current;
+    Pets current;
+
     current.fileExist=0;
 
-    fseek(archive,sizeof (person)*codUser,SEEK_SET);
-    fwrite(&current,sizeof(person),1,archive);
+    fseek(archive,sizeof (Pets)*codPet,SEEK_SET);
+    fwrite(&current,sizeof(Pets),1,archive);
 
-    fseek(archive,sizeof (person)*codUser,SEEK_SET);
-    fread(&current,sizeof(person),1,archive);
+    fseek(archive,sizeof (Pets)*codPet,SEEK_SET);
+    fread(&current,sizeof(Pets),1,archive);
 
     if(current.fileExist==0){
         return 1;
@@ -165,8 +183,17 @@ void listPets(char *file_path){
 
     fseek(chvPet,0,SEEK_SET);
     fread(&codPet,sizeof(int),1,chvPet);
+    if(codPet==0){
+        printf("\n#---------------------------------------------------#");
+        printf("\n|                 NAO HA CADASTROS                  |");
+        printf("\n#---------------------------------------------------#");
 
-    printf("%d",codPet);
+    }else {
+        printf("\n#---------------------------------------------------#");
+        printf("\n|                LISTA DE CADASTROS                 |");
+        printf("\n#---------------------------------------------------#");
+    }
+
     for(int j=0;j<codPet;j++)
     {
         fseek(pets, sizeof(Pets) * j, SEEK_SET);
@@ -188,7 +215,7 @@ void listPets(char *file_path){
 }
 
 void changePetsUI(){
-    int codPet;
+    int codPet,codUser;
     char *Name = malloc(sizeof(int) * sizeMaxNames);
     char *Date = malloc(sizeof(char) * 9);
     char *Type = malloc(sizeof(char) * 50);
@@ -213,11 +240,17 @@ void changePetsUI(){
     fgets(Type, 50, stdin);
     //fgets é pra pegar as strings com espaço, o fflush serve pra tirar todo o lixo da string
 
-    changePets("pets.bin",Name,Date,Type,codPet);
+    printf("\n#-----------------------------------------#");
+    printf("\n| > CODIGO DO DONO DO PET:                |");
+    printf("\n#-----------------------------------------#");
+    printf("\n > ");
+    scanf("%d", &codUser);
+
+    changePets("pets.bin",Name,Date,Type,codUser,codPet);
 
 }
 
-void changePets(char *file_path,char *Name,char *Date,char *Type,int codPet){
+void changePets(char *file_path,char *Name,char *Date,char *Type,int codUser,int codPet){
     Pets *current;
     current=(Pets *)malloc(sizeof(Pets));
 
@@ -230,16 +263,15 @@ void changePets(char *file_path,char *Name,char *Date,char *Type,int codPet){
     strcpy(current->namePet,Name);
     strcpy(current->datePet,Date);
     strcpy(current->typePet,Type);
+    current->codUser=codUser;
     current->codPet=codPet+1;
     current->fileExist=1;
 
     free(Name);
     free(Date);
     free(Type);
-    printf("%d",codPet);
     fseek(archive,sizeof(Pets)*codPet,SEEK_SET);
     fwrite(current,sizeof(Pets),1,archive);
-    printf("%s",current->namePet);
     fclose(archive);
     free(current);
 }
@@ -253,10 +285,6 @@ void searchByPetCode(char *file_path,int codPet) {
     fseek(archive, sizeof(Pets) * codPet, SEEK_SET);
     fread(&listPets, sizeof(Pets), 1, archive);
     if (listPets.fileExist==1) {
-        printf("\n#-----------------------------------------#");
-        printf("\n|          CADASTRO ENCONTRADO            |");
-        printf("\n#-----------------------------------------#");
-
         printf("\n#-----------------------------------------#");
         printf("\n| > Codigo: %.3d                           |", listPets.codPet);
         printf("\n#-----------------------------------------#");
@@ -274,13 +302,12 @@ void searchByPetCode(char *file_path,int codPet) {
     }
 }
 
-void searchByUserCode(char *file_path,int codUser){
-    codUser--;
+int searchByUserCode(char *file_path,int codUser){
+
     FILE *pets;
     FILE *chvPet;
-    int codPet=0;
+    int codPet=0,countPets=0;
     Pets listPets;
-
     pets=fopen(file_path,"rb");
     chvPet=fopen("keyPets.chv","rb");
 
@@ -290,7 +317,7 @@ void searchByUserCode(char *file_path,int codUser){
     for(int j=0;j<codPet;j++) {
         fseek(pets, sizeof(Pets) * j, SEEK_SET);
         fread(&listPets, sizeof(Pets), 1, pets);
-        if (listPets.codUser == codUser) {
+        if (listPets.codUser == codUser && listPets.fileExist==1) {
             printf("\n#-----------------------------------------#");
             printf("\n| > Codigo: %.3d                           |", listPets.codPet);
             printf("\n#-----------------------------------------#");
@@ -298,14 +325,16 @@ void searchByUserCode(char *file_path,int codUser){
             printf(" > Data de nascimento do pet: %s", listPets.datePet);
             printf(" > Tipo do pet: %s", listPets.typePet);
             printf(" > Codigo do usuario: %.3d", listPets.codUser);
-        } else {
-            printf("\n#-----------------------------------------#");
-            printf("\n|           PET NAO ENCONTRADO            |");
-            printf("\n#-----------------------------------------#");
-
-
+            countPets++;
         }
     }
+    if(countPets==0) {
+        printf("\n#-----------------------------------------#");
+        printf("\n|           NENHUM PET ENCONTRADO         |");
+        printf("\n#-----------------------------------------#");
+        return 0;
+    }
+
 
 }
 
@@ -326,28 +355,31 @@ void orderAlfPet(char *file_path) {
 
     fread(&codPet, sizeof(int), 1, chvPet);
     printf("%d", codPet);
-    int auxCod = codPet;
+    int auxCod=codPet;
     Pets allPets[codPet - 1];
     for (int i = 0; i < codPet; i++) {
         fseek(Pet, sizeof(Pets) * i, SEEK_SET);
         fread(&listPets, sizeof(Pets), 1, Pet);
-
         strcpy(allPets[i].namePet, listPets.namePet);
         strcpy(allPets[i].typePet, listPets.typePet);
         strcpy(allPets[i].datePet, listPets.datePet);
         allPets[i].codUser = listPets.codUser;
         allPets[i].codPet = listPets.codPet;
+        allPets[i].fileExist = listPets.fileExist;
+
+
     }
 
     qsort(allPets, codPet, sizeof(Pets), compareNamePet);
-    printf("%d", allPets[0].namePet);
     for (int i = 0; i < auxCod; i++) {
-        printf("\n#-----------------------------------------#");
-        printf("\n| > Codigo: %.3d                           |", allPets[i].codPet);
-        printf("\n#-----------------------------------------#");
-        printf("\n > Nome do pet: %s", allPets[i].namePet);
-        printf(" > Data de nascimento do pet: %s", allPets[i].datePet);
-        printf(" > Tipo do pet: %s", allPets[i].typePet);
-        printf(" > Codigo do usuario: %.3d", allPets[i].codUser);
+        if (allPets[i].fileExist == 1) {
+            printf("\n#-----------------------------------------#");
+            printf("\n| > Codigo: %.3d                           |", allPets[i].codPet);
+            printf("\n#-----------------------------------------#");
+            printf("\n > Nome do pet: %s", allPets[i].namePet);
+            printf(" > Data de nascimento do pet: %s", allPets[i].datePet);
+            printf(" > Tipo do pet: %s", allPets[i].typePet);
+            printf(" > Codigo do usuario: %.3d", allPets[i].codUser);
+        }
     }
 }
